requirements.txt
absl-py==0.9.0
astor==0.8.1
gast==0.3.3
glob3==0.0.1
google-pasta==0.2.0
grpcio==1.29.0
grpcio-tools==1.29.0
h5py==2.10.0
importlib-metadata==1.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.4
Pillow==7.1.2
protobuf==3.12.0
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.46.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
